fix: move "notification" event args types back into `zwave-js` package
